+++ Weather App/Data/Providers/OpenWeatherMap/Services/OWMWeatherService.swift
@@ -29,7 +29,7 @@ class OWMWeatherService: WeatherService {
 
     init() {}
 
-    func fetchForecast(forLat lat: Double, lon: Double) async throws -> Forecast {
+    func fetchForecast(forLat lat: Float, lon: Float) async throws -> Forecast {
         let url = URL(string: "\(baseURL)/weather?lat=\(lat)&lon=\(lon)&appid=\(apiKey)&units=metric")!
         let (data, response) = try await URLSession.shared.data(
             from: url,
